fix: close delete permanently modal on permanently delete success only (contacts and groups)
 - delete contacts-delete.tsx
 - move tests from contacts-delete.test.tsx to use-delete-contacts.test.tsx
